# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GeoPagos Reto Api. Edwin Jose Mogollon Escalona eJme1992",
        "description": "Documentación de la API para la gestión de torneos",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/players/register": {
            "post": {
                "tags": [
                    "Players"
                ],
                "summary": "Register a new player",
                "description": "Registers a new player with the specified details. Requires JWT token for authentication.",
                "operationId": "f0b9f3867ed66bf53bf702ed24abbcce",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "gender": {
                                        "description": "Gender of the player",
                                        "type": "string",
                                        "enum": [
                                            "male",
                                            "female"
                                        ],
                                        "example": "male"
                                    },
                                    "name": {
                                        "description": "Name of the player",
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "ability": {
                                        "description": "Ability of the player (1-100)",
                                        "type": "integer",
                                        "format": "int32",
                                        "maximum": 100,
                                        "minimum": 1,
                                        "example": 75
                                    },
                                    "atributesPlayer": {
                                        "description": "Attributes of the player as key-value pairs",
                                        "properties": {
                                            "strength": {
                                                "description": "Strength attribute of the player",
                                                "type": "number",
                                                "example": 5
                                            },
                                            "speed": {
                                                "description": "Speed attribute of the player",
                                                "type": "number",
                                                "example": 5
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Player registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "description": "Registered player data",
                                            "properties": {
                                                "id": {
                                                    "description": "Player ID",
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "gender": {
                                                    "type": "string",
                                                    "example": "male"
                                                },
                                                "ability": {
                                                    "type": "integer",
                                                    "format": "int32",
                                                    "example": 75
                                                },
                                                "atributesPlayer": {
                                                    "description": "Attributes of the player as key-value pairs",
                                                    "properties": {
                                                        "strength": {
                                                            "type": "number",
                                                            "example": 5
                                                        },
                                                        "speed": {
                                                            "type": "number",
                                                            "example": 5
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "description": "Validation error messages",
                                            "properties": {
                                                "gender": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "The gender field is required."
                                                    ]
                                                },
                                                "name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "The name field is required."
                                                    ]
                                                },
                                                "ability": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "The ability must be an integer between 1 and 100."
                                                    ]
                                                },
                                                "atributesPlayer": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "The atributes player field is required."
                                                    ]
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/players/gender/{gender}": {
            "get": {
                "tags": [
                    "Players"
                ],
                "summary": "Obtiene una lista de jugadores según el género",
                "description": "Obtiene una lista de jugadores según el género especificado. Requiere un token JWT para la autenticación.",
                "operationId": "f716f06b41148f4da18546a022d9e623",
                "parameters": [
                    {
                        "name": "gender",
                        "in": "path",
                        "description": "Género de los jugadores a buscar (ej: 'male', 'female')",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de jugadores encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "John Doe"
                                                    },
                                                    "gender": {
                                                        "type": "string",
                                                        "example": "male"
                                                    },
                                                    "ability": {
                                                        "type": "integer",
                                                        "example": 85
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontraron jugadores",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No players found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Internal server error."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}